✨ <feat> add files for feedback form.

+++ src/main/java/projetb2/overlook_hotel/repository/FeedbackRepository.java
@@ -2,9 +2,8 @@ package projetb2.overlook_hotel.repository;
 
 import org.springframework.stereotype.Repository;
 import org.springframework.data.jpa.repository.JpaRepository;
-
 import projetb2.overlook_hotel.model.Feedback;
 
 @Repository
-public interface FeedbackRepository extends JpaRepository<Feedback, Long> {
+public interface FeedbackRepository extends JpaRepository<Feedback, Integer> {
 }
