+++ src/main/java/projetb2/overlook_hotel/repository/UserNotificationRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import projetb2.overlook_hotel.model.UserNotification;
 
 @Repository
-public interface UserNotificationRepository extends JpaRepository<UserNotification, Long> {
+public interface UserNotificationRepository extends JpaRepository<UserNotification, Integer> {
 
 }
